mkverobj -> emblob
