+++ LICENSE
@@ -1,6 +1,6 @@
 MIT License
 
-Copyright (c) 2023 Ryan M. Lederman <lederman@gmail.com>
+Copyright (c) 2024 Ryan M. Lederman <lederman@gmail.com>
 
 Permission is hereby granted, free of charge, to any person obtaining a copy
 of this software and associated documentation files (the "Software"), to deal
