+++ SECURITY.md
@@ -1,5 +1,5 @@
 <!-- SPDX-License-Identifier: MIT -->
-<!-- Copyright (c) 2018-current Ryan M. Lederman <lederman@gmail.com> -->
+<!-- Copyright (c) 2018-2024 Ryan M. Lederman <lederman@gmail.com> -->
 
 ## Supported Versions
 
@@ -7,4 +7,4 @@ Any version after the first release (1.0) is qualified for vulnerability reporti
 
 ## Reporting a Vulnerability
 
-Please [create a new security vulnerability issue](https://github.com/aremmell/mkverobj/security/advisories/new) and I will review it as quickly as I can.
+Please [create a new security vulnerability issue](https://github.com/aremmell/emblob/security/advisories/new) and I will review it as quickly as I can.
