+++ example/cexample.c
@@ -1,8 +1,9 @@
 /*
- * appstate.hh
+ * cexample.c
  *
  * Author:    Ryan M. Lederman <lederman@gmail.com>
- * Copyright: Copyright (c) 2018-2023
+ * Copyright: Copyright (c) 2018-2024
+ * Version:   0.2.0
  * License:   The MIT License (MIT)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -22,18 +23,14 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#ifndef _MKVEROBJ_APPSTATE_HH_INCLUDED
-#define _MKVEROBJ_APPSTATE_HH_INCLUDED
+#include <stdio.h>
+#include <stdlib.h>
+#include "emblob/version.h"
 
-namespace mkverobj
+int main(void)
 {
-    class app_state
-    {
-    public:
-        bool created_bin_file = false;
-        bool created_asm_file = false;
-        bool created_obj_file = false;
-    };
-} // !namespace mkverobj
-
-#endif // !_MKVEROBJ_APPSTATE_HH_INCLUDED
+    const version_resource* res = get_version_resource();
+    printf("major: %u, minor: %u, patch: %u, suffix: '%s'\n",
+        res->major, res->minor, res->patch, res->suffix);
+    return EXIT_SUCCESS;
+}
