+++ include/emblob/version.h
@@ -2,7 +2,8 @@
  * version.h
  *
  * Author:    Ryan M. Lederman <lederman@gmail.com>
- * Copyright: Copyright (c) 2018-2023
+ * Copyright: Copyright (c) 2018-2024
+ * Version:   0.2.0
  * License:   The MIT License (MIT)
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy of
@@ -22,35 +23,35 @@
  * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
  * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-#ifndef _MKVEROBJ_VERSION_H_INCLUDED
-#define _MKVEROBJ_VERSION_H_INCLUDED
+#ifndef _EMBLOB_VERSION_H_INCLUDED
+#define _EMBLOB_VERSION_H_INCLUDED
 
 #include <stdint.h>
 #include <stddef.h>
 #include <stdalign.h>
 
 #if defined(__cplusplus)
-# define MKVEROBJ_ALIGNAS alignas
-# define MKVEROBJ_EXTERNAL extern "C"
+# define EMBLOB_ALIGNAS alignas
+# define EMBLOB_EXTERNAL extern "C"
 #else
 # if __STDC_VERSION__ < 201710L
-#  define MKVEROBJ_ALIGNAS _Alignas
-# elif __STDC_VERSION__ > 201710L
-#  define MKVEROBJ_ALIGNAS alignas
+#  define EMBLOB_ALIGNAS _Alignas
+# elif __STDC_VERSION__ >= 201710L
+#  define EMBLOB_ALIGNAS alignas
 # endif
-# define MKVEROBJ_EXTERNAL extern
+# define EMBLOB_EXTERNAL extern
 #endif
 
 #if defined(__APPLE__)
-# define MKVEROBJ_VAR version_data
+# define EMBLOB_VAR version_data
 #else
-# define MKVEROBJ_VAR _version_data
+# define EMBLOB_VAR _version_data
 #endif
 
-MKVEROBJ_EXTERNAL uintptr_t MKVEROBJ_VAR;
+EMBLOB_EXTERNAL uintptr_t EMBLOB_VAR;
 
 /** The size, in chars of the ::version_resource.suffix buffer. */
-#define MKVEROBJ_MAX_SUFFIX 256
+#define EMBLOB_MAX_SUFFIX 256
 
 /**
  * @struct version_resource
@@ -59,17 +60,17 @@ MKVEROBJ_EXTERNAL uintptr_t MKVEROBJ_VAR;
  * at runtime by way of the functions exported in this header.
  */
 typedef struct {
-    MKVEROBJ_ALIGNAS(4)
+    EMBLOB_ALIGNAS(4)
 
     uint16_t major; /**< Major version number. */
     uint16_t minor; /**< Minor version number. */
     uint16_t patch; /**< Patch/build/revision number. */
 
     /**
-     * A buffer of ::MKVEROBJ_MAX_SUFFIX chars in size to be used however
+     * A buffer of ::EMBLOB_MAX_SUFFIX chars in size to be used however
      * you see fit.
      */
-    char suffix[MKVEROBJ_MAX_SUFFIX];
+    char suffix[EMBLOB_MAX_SUFFIX];
 } version_resource;
 
 #if defined(__cplusplus)
@@ -78,12 +79,12 @@ typedef struct {
 
 /**
  * Returns a pointer to the version resource struct embedded in the linker
- * input object file generated by mkverobj.
+ * input object file generated by emblob.
  */
 static inline
 const version_resource* get_version_resource(void)
 {
-    return (version_resource*)&MKVEROBJ_VAR;
+    return (version_resource*)&EMBLOB_VAR;
 }
 
 /**
@@ -126,4 +127,4 @@ const char* get_version_suffix(void)
     }
 #endif
 
-#endif // !_MKVEROBJ_VERSION_H_INCLUDED
+#endif // !_EMBLOB_VERSION_H_INCLUDED
